+++ src/server/cf.h
@@ -22,7 +22,7 @@ namespace K {
            argWhitelist    = "";
     protected:
       void load(int argc, char** argv) {
-        cout << BGREEN << "K" << RGREEN << " build " << K_BUILD << " " << K_STAMP << "." << BRED << '\n';
+        screen = new SH();
         static const struct option args[] = {
           {"help",         no_argument,       0,               'h'},
           {"colors",       no_argument,       &argColors,        1},
@@ -97,67 +97,67 @@ namespace K {
               << BGREEN << "K" << RGREEN << " bugkiller: " << RYELLOW << "https://github.com/ctubio/Krypto-trading-bot/issues/new" << '\n'
               << RGREEN << "  downloads: " << RYELLOW << "ssh://git@github.com/ctubio/Krypto-trading-bot" << '\n';
             case '?': cout
-              << FN::uiT() << "Usage:" << BYELLOW << " ./K.sh [arguments]" << '\n'
-              << FN::uiT() << "[arguments]:" << '\n'
-              << FN::uiT() << RWHITE << "-h, --help                - show this help and quit." << '\n'
-              << FN::uiT() << RWHITE << "    --autobot             - automatically start trading on boot." << '\n'
-              << FN::uiT() << RWHITE << "    --dustybot            - do not automatically cancel all orders on exit." << '\n'
-              << FN::uiT() << RWHITE << "    --naked               - do not display CLI, print output to stdout instead." << '\n'
-              << FN::uiT() << RWHITE << "    --headless            - do not listen for UI connections," << '\n'
-              << FN::uiT() << RWHITE << "                            ignores '--without-ssl', '--whitelist' and '--port'." << '\n'
-              << FN::uiT() << RWHITE << "    --without-ssl         - do not use HTTPS for UI connections (use HTTP only)." << '\n'
-              << FN::uiT() << RWHITE << "-L, --whitelist=IP        - set IP or csv of IPs to allow UI connections," << '\n'
-              << FN::uiT() << RWHITE << "                            alien IPs will get a zip-bomb instead." << '\n'
-              << FN::uiT() << RWHITE << "-P, --port=NUMBER         - set NUMBER of an open port to listen for UI connections." << '\n'
-              << FN::uiT() << RWHITE << "-u, --user=WORD           - set allowed WORD as username for UI connections," << '\n'
-              << FN::uiT() << RWHITE << "                            mandatory but may be 'NULL'." << '\n'
-              << FN::uiT() << RWHITE << "-p, --pass=WORD           - set allowed WORD as password for UI connections," << '\n'
-              << FN::uiT() << RWHITE << "                            mandatory but may be 'NULL'." << '\n'
-              << FN::uiT() << RWHITE << "-e, --exchange=NAME       - set exchange NAME for trading, mandatory one of:" << '\n'
-              << FN::uiT() << RWHITE << "                            'COINBASE', 'BITFINEX',  'BITFINEX_MARGIN', 'HITBTC'," << '\n'
-              << FN::uiT() << RWHITE << "                            'OKCOIN', 'OKEX', 'KORBIT', 'POLONIEX' or 'NULL'." << '\n'
-              << FN::uiT() << RWHITE << "-c, --currency=PAIRS      - set currency pairs for trading (use format" << '\n'
-              << FN::uiT() << RWHITE << "                            with '/' separator, like 'BTC/EUR')." << '\n'
-              << FN::uiT() << RWHITE << "-A, --apikey=WORD         - set (never share!) WORD as api key for trading," << '\n'
-              << FN::uiT() << RWHITE << "                            mandatory." << '\n'
-              << FN::uiT() << RWHITE << "-S, --secret=WORD         - set (never share!) WORD as api secret for trading," << '\n'
-              << FN::uiT() << RWHITE << "                            mandatory." << '\n'
-              << FN::uiT() << RWHITE << "-X, --passphrase=WORD     - set (never share!) WORD as api passphrase for trading," << '\n'
-              << FN::uiT() << RWHITE << "                            mandatory but may be 'NULL'." << '\n'
-              << FN::uiT() << RWHITE << "-U, --username=WORD       - set (never share!) WORD as api username for trading," << '\n'
-              << FN::uiT() << RWHITE << "                            mandatory but may be 'NULL'." << '\n'
-              << FN::uiT() << RWHITE << "-H, --http=URL            - set URL of api HTTP/S endpoint for trading," << '\n'
-              << FN::uiT() << RWHITE << "                            mandatory." << '\n'
-              << FN::uiT() << RWHITE << "-W, --wss=URL             - set URL of api SECURE WS endpoint for trading," << '\n'
-              << FN::uiT() << RWHITE << "                            mandatory." << '\n'
-              << FN::uiT() << RWHITE << "-d, --database=PATH       - set alternative PATH to database filename," << '\n'
-              << FN::uiT() << RWHITE << "                            default PATH is '/data/db/K.*.*.*.db'," << '\n'
-              << FN::uiT() << RWHITE << "                            any route to a filename is valid," << '\n'
-              << FN::uiT() << RWHITE << "                            or use ':memory:' (see sqlite.org/inmemorydb.html)." << '\n'
-              << FN::uiT() << RWHITE << "-s, --ewma-short=PRICE    - set initial ewma short value," << '\n'
-              << FN::uiT() << RWHITE << "                            overwrites the value from the database." << '\n'
-              << FN::uiT() << RWHITE << "-m, --ewma-medium=PRICE   - set initial ewma medium value," << '\n'
-              << FN::uiT() << RWHITE << "                            overwrites the value from the database." << '\n'
-              << FN::uiT() << RWHITE << "-l, --ewma-long=PRICE     - set initial ewma long value," << '\n'
-              << FN::uiT() << RWHITE << "                            overwrites the value from the database." << '\n'
-              << FN::uiT() << RWHITE << "-V, --ewma-verylong=PRICE - set initial ewma verylong value," << '\n'
-              << FN::uiT() << RWHITE << "                            overwrites the value from the database." << '\n'
-              << FN::uiT() << RWHITE << "-M, --market-limit=NUMBER - set NUMBER of maximum price levels for the orderbook," << '\n'
-              << FN::uiT() << RWHITE << "                            default NUMBER is '321' and the minimum is '15'." << '\n'
-              << FN::uiT() << RWHITE << "                            locked bots smells like '--market-limit=3' spirit." << '\n'
-              << FN::uiT() << RWHITE << "    --debug-secret        - print (never share!) secret inputs and outputs." << '\n'
-              << FN::uiT() << RWHITE << "    --debug-events        - print detailed output about event handlers." << '\n'
-              << FN::uiT() << RWHITE << "    --debug-orders        - print detailed output about exchange messages." << '\n'
-              << FN::uiT() << RWHITE << "    --debug-quotes        - print detailed output about quoting engine." << '\n'
-              << FN::uiT() << RWHITE << "    --debug-wallet        - print detailed output about target base position." << '\n'
-              << FN::uiT() << RWHITE << "    --debug               - print detailed output about all the (previous) things!" << '\n'
-              << FN::uiT() << RWHITE << "    --colors              - print highlighted output." << '\n'
-              << FN::uiT() << RWHITE << "    --ignore-sun          - do not switch UI to light theme on daylight." << '\n'
-              << FN::uiT() << RWHITE << "    --ignore-moon         - do not switch UI to dark theme on moonlight." << '\n'
-              << FN::uiT() << RWHITE << "-k, --matryoshka=URL      - set Matryoshka link URL of the next UI." << '\n'
-              << FN::uiT() << RWHITE << "-K, --title=WORD          - set WORD as UI title to identify different bots." << '\n'
-              << FN::uiT() << RWHITE << "    --free-version        - work with all market levels but slowdown with 21 XMR hash." << '\n'
-              << FN::uiT() << RWHITE << "-v, --version             - show current build version and quit." << '\n'
+              << ((SH*)screen)->uiT() << "Usage:" << BYELLOW << " ./K.sh [arguments]" << '\n'
+              << ((SH*)screen)->uiT() << "[arguments]:" << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-h, --help                - show this help and quit." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --autobot             - automatically start trading on boot." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --dustybot            - do not automatically cancel all orders on exit." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --naked               - do not display CLI, print output to stdout instead." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --headless            - do not listen for UI connections," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            ignores '--without-ssl', '--whitelist' and '--port'." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --without-ssl         - do not use HTTPS for UI connections (use HTTP only)." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-L, --whitelist=IP        - set IP or csv of IPs to allow UI connections," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            alien IPs will get a zip-bomb instead." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-P, --port=NUMBER         - set NUMBER of an open port to listen for UI connections." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-u, --user=WORD           - set allowed WORD as username for UI connections," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            mandatory but may be 'NULL'." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-p, --pass=WORD           - set allowed WORD as password for UI connections," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            mandatory but may be 'NULL'." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-e, --exchange=NAME       - set exchange NAME for trading, mandatory one of:" << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            'COINBASE', 'BITFINEX',  'BITFINEX_MARGIN', 'HITBTC'," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            'OKCOIN', 'OKEX', 'KORBIT', 'POLONIEX' or 'NULL'." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-c, --currency=PAIRS      - set currency pairs for trading (use format" << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            with '/' separator, like 'BTC/EUR')." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-A, --apikey=WORD         - set (never share!) WORD as api key for trading," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            mandatory." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-S, --secret=WORD         - set (never share!) WORD as api secret for trading," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            mandatory." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-X, --passphrase=WORD     - set (never share!) WORD as api passphrase for trading," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            mandatory but may be 'NULL'." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-U, --username=WORD       - set (never share!) WORD as api username for trading," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            mandatory but may be 'NULL'." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-H, --http=URL            - set URL of api HTTP/S endpoint for trading," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            mandatory." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-W, --wss=URL             - set URL of api SECURE WS endpoint for trading," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            mandatory." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-d, --database=PATH       - set alternative PATH to database filename," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            default PATH is '/data/db/K.*.*.*.db'," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            any route to a filename is valid," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            or use ':memory:' (see sqlite.org/inmemorydb.html)." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-s, --ewma-short=PRICE    - set initial ewma short value," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            overwrites the value from the database." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-m, --ewma-medium=PRICE   - set initial ewma medium value," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            overwrites the value from the database." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-l, --ewma-long=PRICE     - set initial ewma long value," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            overwrites the value from the database." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-V, --ewma-verylong=PRICE - set initial ewma verylong value," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            overwrites the value from the database." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-M, --market-limit=NUMBER - set NUMBER of maximum price levels for the orderbook," << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            default NUMBER is '321' and the minimum is '15'." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "                            locked bots smells like '--market-limit=3' spirit." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --debug-secret        - print (never share!) secret inputs and outputs." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --debug-events        - print detailed output about event handlers." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --debug-orders        - print detailed output about exchange messages." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --debug-quotes        - print detailed output about quoting engine." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --debug-wallet        - print detailed output about target base position." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --debug               - print detailed output about all the (previous) things!" << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --colors              - print highlighted output." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --ignore-sun          - do not switch UI to light theme on daylight." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --ignore-moon         - do not switch UI to dark theme on moonlight." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-k, --matryoshka=URL      - set Matryoshka link URL of the next UI." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-K, --title=WORD          - set WORD as UI title to identify different bots." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "    --free-version        - work with all market levels but slowdown with 21 XMR hash." << '\n'
+              << ((SH*)screen)->uiT() << RWHITE << "-v, --version             - show current build version and quit." << '\n'
               << RGREEN << "  more help: " << RYELLOW << "https://github.com/ctubio/Krypto-trading-bot/blob/master/MANUAL.md" << '\n'
               << BGREEN << "K" << RGREEN << " questions: " << RYELLOW << "irc://irc.domirc.net:6667/##tradingBot" << '\n'
               << RGREEN << "  home page: " << RYELLOW << "https://ca.rles-tub.io./trades" << '\n';
@@ -191,10 +191,12 @@ namespace K {
           argApikey,    argSecret,
           argUsername,  argPassphrase,
           argHttp,      argWss,
-          argMaxLevels, argDebugSecret
+          argMaxLevels, argDebugSecret,
+          ((SH*)screen)->config(
+            argNaked,    argColors,
+            argExchange, argCurrency
+          )
         );
-        if (argNaked) return;
-        FN::screen_config(argColors, argExchange, argCurrency);
       };
     private:
       inline mCoinId base() {
