+++ src/server/km.h
@@ -582,14 +582,15 @@ namespace K {
               RBLUE[]  = "\033[0;34m", RPURPLE[] = "\033[0;35m", RCYAN[]  = "\033[0;36m", RWHITE[]  = "\033[0;37m",
               BBLACK[] = "\033[1;30m", BRED[]    = "\033[1;31m", BGREEN[] = "\033[1;32m", BYELLOW[] = "\033[1;33m",
               BBLUE[]  = "\033[1;34m", BPURPLE[] = "\033[1;35m", BCYAN[]  = "\033[1;36m", BWHITE[]  = "\033[1;37m";
-  static WINDOW *wBorder = nullptr,
-                *wLog    = nullptr;
   static ostringstream             THIS_WAS_A_TRIUMPH;
   static vector<function<void()>*> gwEndings;
+  static function<void()>* shResize;
+  static void *screen = nullptr;
   class Gw {
     public:
       virtual string A() = 0;
-      static Gw *config(mCoinId, mCoinId, string, int, string, string, string, string, string, string, int, int);
+      void *screen = nullptr;
+      static Gw *config(mCoinId, mCoinId, string, int, string, string, string, string, string, string, int, int, void*);
       function<void(mOrder)>        evDataOrder;
       function<void(mTrade)>        evDataTrade;
       function<void(mWallet)>       evDataWallet;
