+++ src/server/og.h
@@ -10,7 +10,7 @@ namespace K {
       void load() {
         for (json &it : ((DB*)memory)->load(mMatter::Trades))
           tradesHistory.push_back(it);
-        FN::log("DB", string("loaded ") + to_string(tradesHistory.size()) + " historical Trades");
+        ((SH*)screen)->log("DB", string("loaded ") + to_string(tradesHistory.size()) + " historical Trades");
       };
       void waitData() {
         gw->evDataOrder = [&](mOrder k) {                           _debugEvent_
@@ -186,7 +186,7 @@ namespace K {
           0, 0, 0, 0, 0, fee, false
         );
         ((EV*)events)->ogTrade(&trade);
-        FN::log(trade, gw->name);
+        ((SH*)screen)->log(trade, gw->name);
         if (qp->_matchPings) {
           mPrice widthPong = qp->widthPercentage
             ? qp->widthPongPercentage * trade.price / 100
@@ -267,7 +267,7 @@ namespace K {
           } else ++it;
       };
       function<void(string)> debug = [&](string k) {
-        FN::log("DEBUG", string("OG ") + k);
+        ((SH*)screen)->log("DEBUG", string("OG ") + k);
       };
   };
 }
